Use a ridiculously large z-index for the Lock
When the Lock appears it should cover all the elements in the page.
That was not the case when there was an element with a z-index greater
than one. This patch is not a silver bullet though, elements with a
z-order larger than ten millions will still show up above the Lock
(but that is unlikely and easy to fix).