Remove ability to reset Lock state when closing it
Right now modes don't need to clean the Lock state when is closed
because it is throwed away, there's no way to reopen the
Lock. However, the possibility to reopen the Lock is still being
discussed and we might need to restore this option, but we can think
about how to that later.