Hide Gravatar when showing an auxiliary pane
We were hiding the Gravatar only when the Lock was done, but that just
a special case of an auxiliary pane showing up. It worked so far
because a Gravatar is never shown along with the select location
auxiliary pane (the only auxiliary that is shown when the Lock is not
done). However, in the future that might not be the case. Also, this
change will allows us to get rid of some code.