Add translations for login/sign up tabs
This introduces the translations 'signin:singinText',
'signup:signinText' and 'signup:signupText'. The first two are used
for the login tab depending on which screen you are on. The last one
along with the existen 'signin:signupText' are used for the sign up
tab.

The design changes in v8 introduced navigation tabs for the login and
sign up screen. The title of the login screen ('signin:title') was
used for the label of the login tab. The label of the signup tab was
the label for the signup button used in the previous design.

That caused a problem when changing the login screen title, because it
also changed the login tab.