Replace "done" callback for "sign in" callback
The "done" semantics are confusing under circustances were the job of
the Lock can be considered done for very different reasons. For
instance, we allow to choose between receiving a magic link via email
and signing in with a social provider. This will cause the "done"
callback to execute in two different situations: a) the email has been
sent or b) the user signed in with a social provider.

To avoid that issue, we will execute the callback only during a
successfull/failed sign in attempt, and in the future we will trigger
an event for rest.

The callback will no longer be invoked in the following situations:
- attempt to sign up
- attempt to send a magic link email
- attempt to resend a magic link email