Remove tabs support from box (almost completely)
Box should have no business with the content (a.k.a. the white part),
otherwise its responsibilities will have no boundaries. For instance, to
display the SSO notice we'll need to add another prop because it needs
to be shown above the tabs.

Box still takes a tabs prop because we need to add an extra class to an
outer div when tabs are rendered so it displays adequately in mobile
devices. It should be removed by something generic in the future.