Replace strategy for social header
Before we were showing a message only when there was no header and the
buttons were small.

Now we have i18n keys for a message that will be shown during
login (`socialLoginInstructions`) and another for sign
up (`socialSignUpInstructions`).

The new approach will not make a distinction between the button size,
but it is easier to understand, more predictable and doesn't force users
to care about the distinction.