Fixes corporate network connection usage. Closes #590
Improves handling of the format sent by the server using and AD connection with Kerberos authentication.

When connecting from a corporate network a user can choose to connect using Kerberos or certificate authentication. In any of these cases lock relies in information provided by the server to validate that the user is in a corporate network and to obtain the name and strategy of the connection associated with the network. 

However, the format of the information is flat inside the `sso` object and not in the connection format used in other cases. The fix manages the flat format and creates a temporary object with the known format to allow `KerberosScreen` to work as expected.