Add resumeAuth method and autoParseHash flag (#790)
When Lock redirects to the redirect url, it puts the params after a # (eg: #id_token=foo&access_token=bar. This causes some issues with client side routers that use # to handle urls. An example is auth0-samples/auth0-angularjs2-systemjs-sample#40.

This PR adds an option to prevent parsing the hash when you instantiate lock (disableParseHashOnInitialization) and adds the resumeAuth method to resume the authentication steps (including all the events)