Added unhandled authorization error
We found an issue with the "last time you logged in with" screen where the silent auth fails because the inactivity timeout/required login after time limit was reached. As if someone went on vacation, came back, and tried to sso back in. The 'login_required' error gets fired because the user needs to re-authenticate but was not emitted as an authorization error. Without this change, there is no way to handle this scenario as it fails to fire lock.on('authorization_error') event.