Disable spellcheck and autocorrect on all sensitive input fields
These were already disabled for the username field spellcheck has recently been discovered to
potentially disclose information to browser vendors that provide this feature. So disabling it
helps to prevent this.

autocorrect is a safari only feature that will autocorrect what it believes to be misspelled words,
so this is disabled for user experience